requirements.txt
Django==4.2.1
django-cleanup==7.0.0
Pillow==9.5.0
razorpay==1.3.0
djangorestframework==3.14.0

# key id-rzp_test_GtCud6v3BP5VGm
# key secret-0VzfjHQNclSACQoJXwrLF9vK
